Moved from unsigned int to int

+++ utils/spgraph.h
@@ -7,11 +7,11 @@ class SpGraph
 
 		float volume;
 		int size;
-		float num_edges;
+		int num_edges;
 		bool directed;
 
-		SpGraph(unsigned int, unsigned int);
-		void add_node(unsigned int, unsigned int*, float*, unsigned int);
+		SpGraph(int, int);
+		void add_node(int, int*, float*, int);
 		
 		void print_degree();
 		void print_graph();
